requirements.txt
fastapi==0.103.1
pydantic==2.3.0
uvicorn==0.23.2
sqlalchemy==2.0.21
aiosmtplib==2.0.2
aioimaplib==1.0.1
python-dotenv==1.0.0
email-validator==2.0.0.post2
jinja2==3.1.2
aiosqlite==0.19.0
fastapi-mcp==0.4.0  # Fast Model-Controller-Processor
python-multipart==0.0.6
httpx==0.24.1
email-reply-parser==0.5.12
schedule==1.2.0
apscheduler==3.10.4
fastapi-utils==0.2.1
python-dateutil==2.8.2
aiohttp==3.8.5